lift watcher to test_utils; trim comments
